Move uefi_platform to library.uefi
