+++ chipsec/library/uefi/uefi_platform.py
