+++ chipsec/modules/tools/uefi/s3script_modify.py
@@ -104,7 +104,7 @@ from chipsec.library.returncode import ModuleResult
 from chipsec.library.logger import print_buffer_bytes
 from chipsec.hal.common.uefi import UEFI
 from chipsec.library.uefi.uefi_common import S3BootScriptOpcode, script_width_values, script_width_formats, op_io_pci_mem, op_dispatch
-from chipsec.hal.common.uefi_platform import encode_s3bootscript_entry, id_s3bootscript_type, create_s3bootscript_entry_buffer
+from chipsec.library.uefi.uefi_platform import encode_s3bootscript_entry, id_s3bootscript_type, create_s3bootscript_entry_buffer
 
 ########################################################################################################
 #
